+++ src/design/plone/contenttypes/behaviors/argomenti.py
@@ -99,6 +99,27 @@ class IArgomentiNews(IArgomentiSchema):
         default=[],
     )
 
+    correlato_in_evidenza = RelationList(
+        title=_("correlato_in_evidenza_label", default="Correlato in evidenza"),
+        description=_(
+            "correlato_in_evidenza_help",
+            default="Seleziona un correlato da mettere in evidenza per questo"
+            " contenuto.",
+        ),
+        value_type=RelationChoice(
+            title=_("Correlato in evidenza"),
+            vocabulary="plone.app.vocabularies.Catalog",
+        ),
+        required=False,
+        default=[],
+    )
+
+    model.fieldset(
+        "correlati",
+        label=_("correlati_label", default="Contenuti collegati"),
+        fields=["correlato_in_evidenza"],
+    )
+
 
 @provider(IFormFieldProvider)
 class IArgomentiEvento(IArgomentiSchema):